+++ src/index.ts
@@ -232,6 +232,7 @@ class Sim800L extends EventEmitter {
         // We define the handler, which will search of an OK end of query
         const handler: JobHandler = (buffer, job) => {
           if (isOk(buffer)) {
+            this.emit('modemready', true);
             this.logger.debug(`checkmodem - modem online`);
             job.callback!({
               uuid: job.uuid,
@@ -247,6 +248,7 @@ class Sim800L extends EventEmitter {
             });
             job.ended = true;
           } else if (getError(buffer).isError) {
+            this.emit('modemready', false);
             this.logger.error(`checkmodem - modem error`);
             job.callback!({
               uuid: job.uuid,
@@ -564,6 +566,7 @@ class Sim800L extends EventEmitter {
             return;
           }
           this.logger.verbose(`checknetwork - status: ${networkStatus}`);
+          this.emit('networkstatus', { networkAction, networkStatus });
           callback({
             uuid: job.uuid,
             type: job.type,
@@ -843,7 +846,6 @@ class Sim800L extends EventEmitter {
     if ([ConnectionStatus.REGISTERED, ConnectionStatus.ROAMING].includes(network.networkStatus) && !this.networkReady) {
       this.logger.warn(`networkhandler - network is now ready`);
       this.networkReady = true;
-      this.emit('networkstatus', network);
     }
     if (
       ![ConnectionStatus.REGISTERED, ConnectionStatus.IN_PROGRESS, ConnectionStatus.ROAMING].includes(
@@ -853,16 +855,13 @@ class Sim800L extends EventEmitter {
     ) {
       this.logger.warn(`networkhandler - network lost, status: ${network.networkStatus}`);
       this.logger.verbose(`networkhandler - ${4 - this.networkRetry} checks remaining before modem reset`);
-      this.emit('networkstatus', network);
       this.networkReady = false;
       this.networkRetry += 1;
     } else {
       if (this.networkReady) {
         this.logger.debug(`networkhandler - network online`);
-        this.emit('networkstatus', network);
         this.networkRetry = 0;
       } else {
-        this.emit('networkstatus', network);
         this.logger.warn(`networkhandler - waiting for network, status: ${network.networkStatus}`);
         this.logger.verbose(`networkhandler - ${4 - this.networkRetry} checks remaining before modem reset`);
         this.networkRetry += 1;